Advanced stats pt 3 taught by Dr. Gu
Advanced Statistical Methods at UCSB. This is the third quarter of the one year graduate course 220ABC on applied statistical methods. The aim is to develop analytical skill for the statistical analysis of data, with emphasis on the basis for the methods, the implementation of the methods, and report writing.